backend/requirements.txt
alembic==1.16.5
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3